+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Change Log
 
+### ? - ?
+
+##### Fixes :wrench:
+
+- Fixed a bug where collision settings were only being applied to the first primitive in a glTF.
+
 ### v1.17.0 - 2022-09-01
 
 ##### Additions :tada: