+++ CHANGES.md
@@ -6,6 +6,7 @@
 
 - Improved the dithered transition between levels-of-detail, making it faster and eliminating depth fighting.
 - Add an option on the `Cesium3DTileset` to change the tileset's mobility. This allows users to make tileset movable at runtime, if needed.  
+- `ACesiumCreditSystem` now has a Blueprint-accessible property for the `CreditsWidget`. This is useful to, for example, move the credits to an in-game billboard rather than a 2D overlay.
 
 ##### Fixes :wrench: